added delete warning dialogue

+++ src/components/ReservationsList.js
@@ -13,7 +13,10 @@ class ReservationsList extends React.Component {
   }
 
   handleClick(res) {
-    this.props.deleteReservationsWithDispatch(res)
+    let r = window.confirm("Are you sure you want to remove this reservation?");
+    if (r == true) {
+      this.props.deleteReservationsWithDispatch(res)
+    }
   }
 
   render() {
